(feat) Refactor to reusable library crate
