+++ src/lib.rs
@@ -1,0 +1,13 @@
+pub mod app;
+pub mod cli;
+pub mod format;
+pub mod model;
+pub mod provider;
+pub mod provider_jira;
+pub mod provider_local;
+pub mod store_fs;
+pub mod ui;
+
+pub use app::{App, Action};
+pub use model::{Board, Card, Column};
+pub use provider::{Provider, ProviderError};
